parker_dycops2022/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.10.0
bleach==4.1.0
bunch==1.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
Flask==2.1.0
Flask-Cors==3.0.10
fonttools==4.31.2
idaes-pse==1.13.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
ipykernel==6.10.0
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.0
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.10
numpy==1.22.3
omlt==0.3.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
Pint==0.18
pluggy==1.0.0
ply==3.11
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
Pygments==2.11.2
Pyomo==6.4.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
rbfopt==4.2.4
requests==2.27.1
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
sympy==1.10.1
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
tinydb==4.7.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.0
widgetsnbextension==3.6.0
xlrd==2.0.1
zipp==3.7.0
